[test] refs #142 - Refactor fsm.c for unit tests
